Split DB merge handler modules
