+++ crates/db-merge/src/merge_handler/recovery.rs
@@ -1,0 +1,177 @@
+use cid::Cid;
+use defra_core::block::{Block, CrdtDelta};
+use defra_core::merge::RecoveredBlockMetadata;
+use storage::corekv::{Key, Store};
+use storage::keys::systemstore::{CollectionKey, CollectionVersionKey};
+
+use super::{DbMergeHandler, MergeError};
+
+impl<S: Store, B: blockstore::Blockstore> DbMergeHandler<S, B> {
+    pub(super) async fn recover_metadata_from_block(
+        &self,
+        cid: &Cid,
+        block_data: &[u8],
+    ) -> Result<Option<RecoveredBlockMetadata>, MergeError> {
+        let block =
+            Block::from_dag_cbor(block_data).map_err(|e| MergeError::BlockDecode(e.to_string()))?;
+
+        // Deltas carry no document identity: recover it from the ownership
+        // index (composites can also derive it from their DAG).
+        let doc_id = match &block.delta {
+            CrdtDelta::Composite(_) => match self.resolve_composite_doc_id(cid, &block).await {
+                Ok(doc_id) => doc_id,
+                // No recoverable identity → treat as unrecoverable metadata;
+                // real infrastructure errors propagate.
+                Err(MergeError::MergeFailed(_)) => return Ok(None),
+                Err(e) => return Err(e),
+            },
+            CrdtDelta::Lww(_) | CrdtDelta::Counter(_) => {
+                match self.resolve_field_block_doc_id(cid).await? {
+                    Some(doc_id) => doc_id,
+                    None => return Ok(None),
+                }
+            }
+            _ => return Ok(None),
+        };
+        let Some(collection_id) = block.delta.schema_version_id().map(ToString::to_string) else {
+            return Ok(None);
+        };
+
+        let Some(creator) = self.verify_block_signature(cid, &block, block_data).await? else {
+            return Ok(None);
+        };
+
+        Ok(Some(
+            RecoveredBlockMetadata::new(doc_id, collection_id, creator.clone())
+                .with_verified_creator(Some(creator)),
+        ))
+    }
+
+    /// Look up a previous CollectionVersion from block heads.
+    ///
+    /// For patched collection versions, the block's heads point to previous version CIDs.
+    /// First checks systemstore, then falls back to decoding the head block directly
+    /// from blockstore (for the UnknownCollection case where the initial version
+    /// hasn't been processed yet).
+    pub(crate) async fn resolve_previous_collection_version(
+        &self,
+        block: &Block,
+    ) -> Result<Option<schema::CollectionVersion>, MergeError> {
+        let heads = match &block.heads {
+            Some(heads) if !heads.is_empty() => heads,
+            _ => return Ok(None),
+        };
+
+        for head_cid in heads {
+            // Fast path: check systemstore (KnownCollection case)
+            let head_key = CollectionKey::new(head_cid.to_string());
+            let txn = self.db.new_txn(true).await.map_err(MergeError::Database)?;
+            let systemstore = txn.systemstore().map_err(MergeError::Database)?;
+
+            if let Ok(Some(data)) = systemstore.get(&head_key.bytes()).await {
+                if let Ok(prev) = serde_json::from_slice::<schema::CollectionVersion>(&data) {
+                    tracing::debug!(
+                        head_cid = %head_cid,
+                        name = %prev.name,
+                        collection_id = %prev.collection_id,
+                        "Resolved previous collection version from systemstore"
+                    );
+                    return Ok(Some(prev));
+                }
+            }
+
+            // Slow path: decode head block directly from blockstore.
+            // Build a CollectionVersion from the raw block data without going
+            // through the full merge handler (avoids async recursion).
+            if let Ok(Some(head_block_data)) = self.blockstore.get(head_cid).await {
+                let head_block = Block::from_dag_cbor(&head_block_data).map_err(|e| {
+                    MergeError::BlockDecode(format!("Failed to decode head block: {}", e))
+                })?;
+
+                if let CrdtDelta::CollectionDefinition(head_payload) = &head_block.delta {
+                    if let Some(name) = &head_payload.name {
+                        tracing::debug!(
+                            head_cid = %head_cid,
+                            name = %name,
+                            "Resolved previous collection version from blockstore"
+                        );
+
+                        // Decode field blocks from the head block's links
+                        let mut prev_fields = Vec::new();
+                        if let Some(links) = &head_block.links {
+                            for link in links.iter() {
+                                let field_cid = &link.link;
+                                if let Ok(Some(field_bytes)) = self.blockstore.get(field_cid).await
+                                {
+                                    if let Ok(field_block) = Block::from_dag_cbor(&field_bytes) {
+                                        if let CrdtDelta::FieldDefinition(fp) = &field_block.delta {
+                                            if let Ok(fd) = self.field_definition_to_description(
+                                                fp,
+                                                &field_cid.to_string(),
+                                            ) {
+                                                prev_fields.push(fd);
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+
+                        let head_version_id = head_cid.to_string();
+                        let mut prev = schema::CollectionVersion::new(
+                            name,
+                            &head_version_id,
+                            &head_version_id,
+                            prev_fields,
+                        );
+                        prev.is_active = false;
+                        prev.is_materialized = true;
+
+                        // Ensure _docID is first in fields
+                        if let Some(pos) = prev.fields.iter().position(|f| f.name == "_docID") {
+                            if pos > 0 {
+                                let f = prev.fields.remove(pos);
+                                prev.fields.insert(0, f);
+                            }
+                        }
+
+                        // Store in systemstore so GetCollections can find it
+                        let txn2 = self.db.new_txn(false).await.map_err(MergeError::Database)?;
+                        {
+                            let ss = txn2.systemstore().map_err(MergeError::Database)?;
+                            let key = CollectionKey::new(&head_version_id);
+                            let data = serde_json::to_vec(&prev).map_err(|e| {
+                                MergeError::Storage(format!(
+                                    "Failed to serialize prev collection: {}",
+                                    e
+                                ))
+                            })?;
+                            ss.set(&key.bytes(), &data).await.map_err(|e| {
+                                MergeError::Storage(format!(
+                                    "Failed to store prev collection: {}",
+                                    e
+                                ))
+                            })?;
+                            let vkey =
+                                CollectionVersionKey::new(&head_version_id, &head_version_id);
+                            ss.set(&vkey.bytes(), b"1").await.map_err(|e| {
+                                MergeError::Storage(format!(
+                                    "Failed to store prev version index: {}",
+                                    e
+                                ))
+                            })?;
+                        }
+                        txn2.commit().await.map_err(MergeError::Database)?;
+                        self.db
+                            .add_collection_to_cache(prev.clone())
+                            .map_err(MergeError::Database)?;
+
+                        return Ok(Some(prev));
+                    }
+                }
+            }
+        }
+
+        Ok(None)
+    }
+}
